[REFACTOR] Moved preference keys and category util to core-lib layer.
