+++ core-lib/src/main/java/info/hossainkhan/android/core/data/CategoryNameResolver.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package info.hossainkhan.android.core.model;
+package info.hossainkhan.android.core.data;
 
 import android.support.annotation.StringRes;
