+++ core-lib/src/main/res/values/donottranslate.xml
