+++ mobile/src/main/java/info/hossainkhan/dailynewsheadlines/MainActivity.java
@@ -13,7 +13,7 @@ import info.hossainkhan.android.core.base.BaseActivity;
 import info.hossainkhan.android.core.headlines.HeadlinesContract;
 import info.hossainkhan.android.core.headlines.HeadlinesPresenter;
 import info.hossainkhan.android.core.model.CardItem;
-import info.hossainkhan.android.core.model.CategoryNameResolver;
+import info.hossainkhan.android.core.data.CategoryNameResolver;
 import info.hossainkhan.android.core.model.NavigationRow;
 import timber.log.Timber;
